'precision' type is int.

+++ src/logging.C
@@ -162,8 +162,8 @@ update_lastlog (const char *fname, const char *pty, const char *host)
   if (S_ISDIR (st.st_mode))
     {
       sprintf (lastlogfile, "%.*s/%.*s",
-               sizeof (lastlogfile) - sizeof (pwent->pw_name) - 2, fname,
-               sizeof (pwent->pw_name),
+               (int)(sizeof (lastlogfile) - sizeof (pwent->pw_name) - 2), fname,
+               (int)sizeof (pwent->pw_name),
                (!pwent->pw_name || pwent->pw_name[0] == '\0') ? "unknown"
                : pwent->pw_name);
       if ((fd = open (lastlogfile, O_WRONLY | O_CREAT, 0644)) >= 0)
